Moved new_eyewitness to EyeWitness.py, fixed relative dirs to work with cwd
