--- fast.txt
@@ -1,10 +1,0 @@
-www.google.com
-www.google.co.uk
-www.google.co.in
-www.example.org
-www.google.co.jp
-www.google.at
-www.google.com.au
-www.google.com.hk
-www.google.com.jm
-www.google.co.kr
