--- sites.txt
@@ -1,3 +1,0 @@
-www.google.com
-www.reddit.com
-www.gmail.com
