--- top10.txt
@@ -1,10 +1,0 @@
-www.google.com
-www.facebook.com
-www.youtube.com
-www.yahoo.com
-www.baidu.com
-www.amazon.com
-www.wikipedia.org
-www.taobao.com
-www.twitter.com
-www.google.co.uk
