--- top100.txt
@@ -1,100 +1,0 @@
-360buy.com
-google.com
-facebook.com
-youtube.com
-yahoo.com
-baidu.com
-wikipedia.org
-live.com
-qq.com
-amazon.com
-taobao.com
-twitter.com
-blogspot.com
-linkedin.com
-google.co.in
-yahoo.co.jp
-bing.com
-sina.com.cn
-yandex.ru
-msn.com
-ebay.com
-wordpress.com
-google.com.hk
-google.co.jp
-vk.com
-google.de
-163.com
-google.co.uk
-weibo.com
-google.fr
-microsoft.com
-mail.ru
-googleusercontent.com
-tumblr.com
-pinterest.com
-google.com.br
-fc2.com
-google.ru
-xvideos.com
-conduit.com
-hao123.com
-ask.com
-paypal.com
-amazon.co.jp
-craigslist.org
-google.it
-xhamster.com
-google.es
-babylon.com
-blogger.com
-tmall.com
-apple.com
-imdb.com
-sohu.com
-avg.com
-bbc.co.uk
-google.com.mx
-amazon.de
-soso.com
-odnoklassniki.ru
-mywebsearch.com
-instagram.com
-google.ca
-go.com
-pornhub.com
-ifeng.com
-aol.com
-alibaba.com
-stackoverflow.com
-youku.com
-t.co
-adcash.com
-360.cn
-adobe.com
-adf.ly
-thepiratebay.se
-rakuten.co.jp
-blogspot.in
-bp.blogspot.com
-flickr.com
-about.com
-redtube.com
-ebay.de
-alipay.com
-cnn.com
-google.com.au
-livedoor.com
-neobux.com
-imgur.com
-sogou.com
-amazon.co.uk
-google.com.tr
-google.pl
-uol.com.br
-dailymotion.com
-netflix.com
-huffingtonpost.com
-xnxx.com
-livejasmin.com
-ebay.co.uk
